replace logic in class
